temp

+++ src/jitlayers.h
@@ -202,7 +202,7 @@ struct jl_returninfo_t {
     size_t union_bytes;
     size_t union_align;
     size_t union_minalign;
-    unsigned return_roots;
+    size_t return_roots; // TODO: rename
 };
 
 struct jl_codegen_call_target_t {
